+++ example/lib/basic_example.g.dart
@@ -6,7 +6,67 @@ part of 'basic_example.dart';
 // DataClassGenerator
 // **************************************************************************
 
-// ignore_for_file: annotate_overrides
+// ignore_for_file: annotate_overrides, unused_element
+
+mixin _$Order {
+  Order get _self => this as Order;
+
+  Iterable<Object?> get _props sync* {
+    yield _self.product;
+  }
+
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is _$Order &&
+          runtimeType == other.runtimeType &&
+          DataClass.$equals(_props, other._props);
+
+  int get hashCode => Object.hashAll(_props);
+
+  String toString() =>
+      (ClassToString('Order')..add('product', _self.product)).toString();
+
+  Order copyWith({
+    Product? product,
+  }) {
+    return Order(
+      product: product ?? _self.product,
+    );
+  }
+
+  Order change(void Function(_OrderChanges c) updates) =>
+      (_OrderChanges._(_self)..update(updates)).build();
+
+  _OrderChanges toChanges() => _OrderChanges._(_self);
+}
+
+class _OrderChanges {
+  late Product product;
+
+  _OrderChanges._(Order dataClass) {
+    replace(dataClass);
+  }
+
+  void update(void Function(_OrderChanges c) updates) => updates(this);
+
+  void replace(covariant Order dataClass) {
+    product = dataClass.product;
+  }
+
+  Order build() => Order(
+        product: product,
+      );
+}
+
+class OrderFields {
+  final String _path;
+
+  const OrderFields([this._path = '']);
+
+  ProductFields get product => ProductFields('${_path}product.');
+
+  String toString() => _path.isEmpty ? 'OrderFields()' : _path;
+}
 
 mixin _$Product {
   Product get _self => this as Product;
@@ -71,4 +131,16 @@ class _ProductChanges {
         title: title,
         extraData: extraData,
       );
+}
+
+class ProductFields {
+  final String _path;
+
+  const ProductFields([this._path = '']);
+
+  String get id => '${_path}id';
+  String get title => '${_path}title';
+  String get extraData => '${_path}extraData';
+
+  String toString() => _path.isEmpty ? 'ProductFields()' : _path;
 }